+++ packages/jsx-flow/src/index.ts
@@ -8,17 +8,15 @@ export { Deferred } from "./components/Deferred.js";
 export { Patch } from "./components/Patch.js";
 export { Generator } from "./components/Generator.js";
 export { Slot, type SlotProps } from "./components/Slot.js";
-export { assertFragmentId } from "./fragmentId.js";
 export type { PatchProps } from "./components/Patch.js";
 export type { GeneratorProps } from "./components/Generator.js";
-export { initFlow, Flow } from "./context.js";
+export { Flow } from "./context.js";
 export type {
   FragmentEffect,
   StreamEffect,
   FlowContext,
   Config,
 } from "./context.js";
-export { streamFlow } from "./streamFlow.js";
 export {
   renderToFlowEvents,
   renderToReadableStream,